cctvai-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
python-multipart
python-dotenv
